+++ src/main/ImGuiManager.cpp
@@ -1,6 +1,6 @@
 /*
  * MacroQuest: The extension platform for EverQuest
- * Copyright (C) 2002-2022 MacroQuest Authors
+ * Copyright (C) 2002-2023 MacroQuest Authors
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License, version 2, as published by
@@ -28,6 +28,7 @@
 
 // move to mq/utils? or do we need a new static lib?
 #include "common/HotKeys.h"
+#include <mq/base/WString.h>
 #include <mq/utils/Benchmarks.h>
 
 #include <cfenv>